+++ packages/button/src/index.ts
@@ -1,1 +1,3 @@
-export * from './Button.js'
+export * from './button/index.js'
+export * from './buttonIcon/index.js'
+export * from './types.js'