+++ packages/button/src/types.ts
@@ -1,0 +1,30 @@
+export const ButtonTypeType = {
+  button: true,
+  submit: true,
+  reset: true,
+}
+
+export const ButtonPaddingType = {
+  sm: true,
+  md: true,
+  lg: true,
+}
+
+export const ButtonTextTransformType = {
+  capitalize: true,
+  uppercase: true,
+  default: true,
+  none: true,
+}
+
+export const ButtonEffectType = {
+  globalScale: true,
+  boxScale: true,
+  opacity: true,
+}
+
+export const ButtonGapType = {
+  sm: true,
+  md: true,
+  lg: true,
+}