+++ packages/styles/src/sizes.ts
@@ -1,3 +1,13 @@
-export const Sizes = {}
+export const Sizes = {
+  xxxs: 1,
+  xxs: 2,
+  xs: 4,
+  sm: 8,
+  md: 16,
+  lg: 32,
+  xl: 64,
+  xxl: 128,
+  xxxl: 256,
+}
 
 export type SizesType = keyof typeof Sizes