+++ apps/storybook/CHANGELOG.md
@@ -6,10 +6,9 @@
 
 - Updated dependencies
   - @pikas-ui/button@1.0.10
+  - @pikas-ui/styles@1.0.10
   - @pikas-ui/icons@1.0.10
   - @pikas-ui/loader@1.0.10
-  - @pikas-ui/separator@1.0.10
-  - @pikas-ui/styles@1.0.10
 
 ## 1.0.9