+++ packages/button/CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ### Patch Changes
 
-- Add separator
+- Add ButtonIcon
 - Updated dependencies
+  - @pikas-ui/styles@1.0.10
   - @pikas-ui/icons@1.0.10
   - @pikas-ui/loader@1.0.10
-  - @pikas-ui/styles@1.0.10
 
 ## 1.0.9